Migrate proto generation to `buf`
- Migrate proto generation to `buf`
- Update linter configuration
- Fix linting issues

+++ broker/machinebroker/server/machine_volume_attach.go
@@ -49,7 +49,7 @@ func (s *Server) getIronCoreVolumeConfig(volume *iri.Volume) (*IronCoreVolumeCon
 	case volume.EmptyDisk != nil:
 		var sizeLimit *resource.Quantity
 		if sizeBytes := volume.EmptyDisk.SizeBytes; sizeBytes > 0 {
-			sizeLimit = resource.NewQuantity(int64(sizeBytes), resource.DecimalSI)
+			sizeLimit = resource.NewQuantity(sizeBytes, resource.DecimalSI)
 		}
 		emptyDisk = &IronCoreVolumeEmptyDiskConfig{
 			SizeLimit: sizeLimit,
